% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_final_table.R
\name{X_create_final_table}
\alias{X_create_final_table}
\title{Function to Create the Final Summary Table}
\usage{
X_create_final_table(x, sink.dir, table.dir, table.name)
}
\arguments{
\item{x}{the table object.}

\item{sink.dir}{(what is this directory?)}

\item{table.dir}{the path to the directory that stores the produced table.}

\item{table.name}{the name of the table (a character string).}
}
\description{
Function to Create the Final Summary Table
}
