% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_final_table.R
\name{X_prep_final_table}
\alias{X_prep_final_table}
\title{Preparing The Data Table for Making The Final Summary Table}
\usage{
X_prep_final_table(x, title.txt, subtitle.txt, stub.txt)
}
\arguments{
\item{x}{a table.}

\item{title.txt}{the table title (a character string).}

\item{subtitle.txt}{the table subtitle (a character string).}

\item{stud.txt}{a stud text (a character string).}
}
\description{
This function takes a table, a title, a subtitle, and a stud text, and puts it into the format required by X_create_final_table.
}
